% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FstBattia.R
\name{Compute_Denominator}
\alias{Compute_Denominator}
\title{Computation of the numerator of the moment estimator}
\usage{
Compute_Denominator(p1, p2)
}
\arguments{
\item{p1}{numeric, frequencies in population 1}

\item{p2}{numeric, frequencies in population 2}
}
\value{
a vector with the denominators of the Fst moment estimator
}
\description{
Computation of the numerator of the moment estimator
}
