% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FstBattia.R
\name{Compute_Nominator}
\alias{Compute_Nominator}
\title{Computation of the numerator of the moment estimator}
\usage{
Compute_Nominator(p1, p2, n1, n2)
}
\arguments{
\item{p1}{numeric, frequencies in population 1}

\item{p2}{numeric, frequencies in population 2}

\item{n1}{numeric, number of gametes in population 1}

\item{n2}{numeric, number of gametes in population 2}
}
\value{
a vector with the  numerators of the Fst moment estimator
}
\description{
Computation of the numerator of the moment estimator
}
