% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FstBattia.R
\name{DF4Plot1Prof}
\alias{DF4Plot1Prof}
\title{DF4Plot1Prof}
\usage{
DF4Plot1Prof(Info, HF, FstProf, Coord = NULL, Threshold = NULL)
}
\arguments{
\item{Info}{a data.frame providing information about markers}

\item{HF}{a data.frame with 2 columns Fst and Weight, as obtained from the \code{HudsonFst.m} function}

\item{FstProf}{an Fst profile, as obtained from the \code{HudsonFst.prof} function}

\item{Coord}{a vector with the minimum and maximum coordinates (i.e. positions along the genome), 
providing the range of the genomic region that will be plotted, optional.}

\item{Threshold}{a numeric value. Markers belonging to regions whose Fst profile is higher than threshold
will be highlighted. Optional.}
}
\value{
a data.frame that can be used as an input for function \code{Plot1Prof}
}
\description{
Shape the data for Fst profile representation
}
