% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FstBattia.R
\name{Freq.filt}
\alias{Freq.filt}
\title{Filtering markers based on allelic frequencies}
\usage{
Freq.filt(FreqNbG, Maf = 0)
}
\arguments{
\item{FreqNbG}{a list of data.frames (one per population) with 2 columns: Freq and NbGamete}

\item{Maf}{a numerci value for the thresholding of minor allelic frequencies}
}
\value{
a vector of positions to be removed
}
\description{
Filtering markers based on allelic frequencies
}
