% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{Freq}
\alias{Freq}
\title{Frequencies of the American and European HGDP populations}
\format{
A data.frame
}
\usage{
Freq
}
\description{
\code{Freq} is a data.frame containing the frequencies of 49,636 markers located on chromosome 1, 
for the 13 American and European populations described in the Stanford HGDP dataset.
}
\keyword{datasets}
