% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fst4pg-package.R
\docType{package}
\name{fst4pg-package}
\alias{fst4pg}
\alias{fst4pg-package}
\title{fst4pg: Genetic Distance Segmentation for Population Genetics}
\description{
Provides efficient methods to compute local and genome wide genetic distances (corresponding to the so called Hudson Fst parameters) through moment method, perform chromosome segmentation into homogeneous Fst genomic regions, and selection sweep detection for multi-population comparison. When multiple profile segmentation is required, the procedure can be parallelized using the future package.
}
\author{
\strong{Maintainer}: Tristan Mary-Huard \email{tristan.mary-huard@agroparistech.fr} (\href{https://orcid.org/0000-0002-3839-9067}{ORCID})

Authors:
\itemize{
  \item Guillem Rigaill \email{guillem.rigaill@inrae.fr} (\href{https://orcid.org/0000-0002-7176-7511}{ORCID})
}

}
\keyword{internal}
