% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FstBattia.R
\name{HudsonFst.gw}
\alias{HudsonFst.gw}
\title{HudsonFst.gw}
\usage{
HudsonFst.gw(HFst.m, Mat = TRUE)
}
\arguments{
\item{HFst.m}{a list of data.frames as obtained with function \code{HudsonFst.m}}

\item{Mat}{boolean, should the result be output as a matrix.}
}
\value{
By default a matrix of Hudson Fst coefficients, a vector otherwise.
}
\description{
Compute genome-wide Hudson Fst moment estimator
}
\examples{
data(Freq);data(NbGamete)
FreqNbG <- BuildFreqNbG(Freq,NbGamete)
HFst.m <- HudsonFst.m(FreqNbG)
HFst.chr <- HudsonFst.gw(HFst.m)
}
