% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FstBattia.R
\name{HudsonFst.m}
\alias{HudsonFst.m}
\title{HudsonFst.m}
\usage{
HudsonFst.m(FreqNbG)
}
\arguments{
\item{FreqNbG}{a list of data.frames (one per population) with 2 columns: Freq and NbGamete}
}
\value{
a list of data.frames with 2 columns: Fst and Weight.
}
\description{
Compute Hudson Fst moment estimator at marker level
}
\examples{
## Load the FreqNbG object build from the HGDP data
data(Freq);data(NbGamete)
FreqNbG <- BuildFreqNbG(Freq,NbGamete)
HFst.m <- HudsonFst.m(FreqNbG)
}
