% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{Info}
\alias{Info}
\title{Marker information}
\format{
A data.frame
}
\usage{
Info
}
\description{
\code{Info} is a data.frame describing the markers located on chromosome 1 in the Stanford 
HGDP dataset. Each row corresponds to a marker, and the 5 columns provide information about
the marker name, its chromosome membership, its position, and its reference and alternative alleles.
}
\keyword{datasets}
