% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{MakeProfile.op}
\alias{MakeProfile.op}
\title{MakeProfile.op}
\usage{
MakeProfile.op(DF, coef.pen.value = 1, sd.y = NULL)
}
\arguments{
\item{DF}{a data.frame with two columns, Fst and Weights, as provided by the \code{HudsonFst.m} function}

\item{coef.pen.value}{coef to use for penaly 2*coef.pen.value*log(n)}

\item{sd.y}{a numeric value corresponding to the (estimated) standard deviation of the signal. If NULL (default) the value is automatically estimated.}
}
\value{
a smoothed profile
}
\description{
Perform segmentation on a given dataset and returns the segmented profile
}
