% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{MakeProfile.sn_nomemory}
\alias{MakeProfile.sn_nomemory}
\title{MakeProfile.sn_nomemory}
\usage{
MakeProfile.sn_nomemory(DF, Kmax, method = "biggest.S3IB", sd.y = NULL)
}
\arguments{
\item{DF}{a data.frame with two columns, Fst and Weights, as provided by the \code{HudsonFst.m} function}

\item{Kmax}{max number of changes used for model selection (check with crops)}

\item{method}{a string, the name of the criterion used for model selection: (1) "givenVariance" = using the penalty of Lebarbier 2005 given a estimator of the variance, (2) "biggest.S3IB" = biggest=TRUE in saut taken from S3IB, (3) "notbiggest.S3IB"
To be chosen amongs3ib.jump, s3ib.nojump, pre, ddse (capushe) or jump (capushe)}

\item{sd.y}{a numeric value corresponding to the (estimated) standard deviation of the signal. If NULL (default) the value is automatically estimated.}
}
\value{
a smoothed profile
}
\description{
Perform segmentation on a given dataset and returns the segmented profile
}
