% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FstBattia.R
\name{MergeRegion}
\alias{MergeRegion}
\title{MergeRegion}
\usage{
MergeRegion(DT, Crit)
}
\arguments{
\item{DT}{a data.frame}

\item{Crit}{a string corresponding to the name of the criterion used for selecting the top regions}
}
\value{
a simplified data.frame (tibble)
}
\description{
Merge adjacent top regions
}
