% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{NbGamete}
\alias{NbGamete}
\title{Number of gametes of the American and European HGDP populations}
\format{
A data.frame
}
\usage{
NbGamete
}
\description{
\code{NbGamete} is a data.frame containing the number of gametes collected 
for the 13 American and European populations described in the Stanford HGDP dataset.
}
\keyword{datasets}
