% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graphs.R
\name{Plot1Prof}
\alias{Plot1Prof}
\title{Plot1Prof}
\usage{
Plot1Prof(DF, Title = "", Range = NULL)
}
\arguments{
\item{DF}{a data.frame, as provided by function \code{DF4Plot1Prof}}

\item{Title}{a string providing a title for the graph, optional.}

\item{Range}{a vector with the minimum and maximum values for the y-axis}
}
\value{
a ggplot object
}
\description{
Display the graphical representation of an Fst profile
}
