% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FstBattia.R
\name{Ratio_Average}
\alias{Ratio_Average}
\title{Computation of the Fst moment estimator}
\usage{
Ratio_Average(Nominator, Denominator)
}
\arguments{
\item{Nominator}{numeric, numerators of the Fst moment estimator}

\item{Denominator}{numeric, denominators of the Fst moment estimator}
}
\value{
a vector with the global Fst estimator
}
\description{
Computation of the Fst moment estimator
}
