% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graphs.R
\name{RawPlot}
\alias{RawPlot}
\title{RawPlot}
\usage{
RawPlot(Info, HF, Coord = NULL, Title = "")
}
\arguments{
\item{Info}{a data.frame providing information about markers}

\item{HF}{a data.frame with 2 columns, Fst and Weight, as provided by the \code{HudsonFst.m} function}

\item{Coord}{a vector with the minimum and maximum coordinates (i.e. positions along the genome), 
providing the range of the genomic region that will be plotted.}

\item{Title}{a string providing a title for the graph.}
}
\value{
a ggplot object.
}
\description{
Plot the Fst estimates along a (portion of) chromosome
}
