% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FstBattia.R
\name{Summarise1Profile}
\alias{Summarise1Profile}
\title{Summarise1Profile}
\usage{
Summarise1Profile(profile, snpinfo)
}
\arguments{
\item{profile}{a vector, corresponding to the Fst profile of a pair of populations}

\item{snpinfo}{a data.frame, providing information about markers}
}
\value{
a data.frame that combines both objects
}
\description{
Summarise1Profile
}
